+++ src/pages/Testimonials.jsx
@@ -3,7 +3,7 @@ import Footer from "../layout/Footer";
 import Section from "../layout/Section";
 import style from "../style";
 import { fullQuotes } from "../data/quotes";
-import { allOrgs } from "data/organisations";
+import { allOrgs } from "data/organizations";
 import useDisplayCategory from "../hooks/useDisplayCategory";
 
 export default function Testimonials() {
