+++ src/pages/home/HomeQuoteCarousel.jsx
@@ -2,7 +2,7 @@ import style from "../../style/index.js";
 import useDisplayCategory from "../../hooks/useDisplayCategory.js";
 import { useState, useEffect } from "react";
 import { quoteData } from "../../data/quotes.js";
-import { orgData } from "../../data/organisations.js";
+import { orgData } from "../../data/organizations.js";
 import Carousel from "../../layout/Carousel.jsx";
 import useCountryId from "../../hooks/useCountryId.js";
 import { Link } from "react-router-dom";
